requirements.txt
manticore[native]
protobuf==3.20
web3
py-solc-x
pydantic
openai
Flask
flask-cors
mythril
